add basic static ui
