+++ lib/auth.js
@@ -18,7 +18,7 @@ module.exports = function bot (fastify, opts, next) {
       }
     },
     // callbackUriParams: { permissions: fastify.config.DISCORD_PERMISSION }, // only for scope=bot
-    scope: ['email', 'identify', 'gdm.join', 'connections'],
+    scope: ['identify'],
     // register a fastify url to start the redirect flow
     startRedirectPath: '/discord',
     // facebook redirect here after the user login
