+++ static/index.html
@@ -1,0 +1,75 @@
+<!doctype html>
+<html lang="en">
+
+<head>
+  <!-- Required meta tags -->
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+
+  <!-- Bootstrap CSS -->
+  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css"
+    integrity="sha384-JcKb8q3iqJ61gNV9KGb8thSsNjpSL0n8PARn9HuZOnIxN0hoP+VmmDGMN5t9UJ0Z" crossorigin="anonymous">
+
+  <script src="https://kit.fontawesome.com/e349d2433a.js" crossorigin="anonymous"></script>
+  <style>
+    body {
+      margin: 2% 10%;
+      background-color: #ededed;
+    }
+
+    .discord {
+      background-color: #7289da;
+      color: #ffffff;
+    }
+  </style>
+  <title>Discord Zodiac!</title>
+</head>
+
+<body>
+  <div class="card text-center">
+    <img src="https://via.placeholder.com/468x120/23272a/99aab5?text=Placeholder" class="card-img-top" alt="...">
+    <div class="card-body">
+      <h5 class="card-title">Hi Guest!</h5>
+      <p class="card-text">Set your Zodiac as your status on Discord!</p>
+      <p>
+        <select class="custom-select custom-select-sm">
+          <option selected>Select your Zodiac Sign</option>
+          <option value="♑ Capricorn">♑ Capricorn</option>
+          <option value="♒ Aquarius">♒ Aquarius</option>
+          <option value="♓ Pisces">♓ Pisces</option>
+          <option value="♈ Aries">♈ Aries</option>
+          <option value="♉ Taurus">♉ Taurus</option>
+          <option value="♊ Gemini">♊ Gemini</option>
+          <option value="♋ Cancer">♋ Cancer</option>
+          <option value="♌ Leo">♌ Leo</option>
+          <option value="♍ Virgo">♍ Virgo</option>
+          <option value="♎ Libra">♎ Libra</option>
+          <option value="♏ Scorpius">♏ Scorpius</option>
+          <option value="♐ Sagittarius">♐ Sagittarius</option>
+          <!--
+          ♑ Capricorn — December 22 - January 20
+          ♒ Aquarius — January 21 - February 19
+          ♓ Pisces — February 20 - March 20
+          ♈ Aries — March 21 - April 20
+          ♉ Taurus — April 21 - May 21
+          ♊ Gemini — May 22 - June 21
+          ♋ Cancer — June 22 - July 22
+          ♌ Leo — July 23 - August 22
+          ♍ Virgo — August 23 - September 23
+          ♎ Libra — September 24 - October 23
+          ♏ Scorpius — October 24 - November 22
+          ♐ Sagittarius — November 23 - December 21
+          -->
+        </select>
+      </p>
+      <p>
+        <a href="/auth/discord" class="btn discord"><i class="fab fa-discord"></i> Login with Discord</a>
+      </p>
+    </div>
+    <div class="card-footer text-muted">
+      Demo <a href="https://github.com/Eomm/fastify-discord-bot-demo" target="_blank">repository</a>
+    </div>
+  </div>
+</body>
+
+</html>
